part3/webapp/final/pypi_web_mongodb_f/requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
beautifulsoup4==4.6.0
Chameleon==3.3
coverage==4.5.1
hupper==1.3
Logbook==1.4.0
Mako==1.0.7
MarkupSafe==1.0
mongoengine==0.15.0
more-itertools==4.2.0
passlib==1.7.1
PasteDeploy==1.5.2
plaster==1.0
plaster-pastedeploy==0.5
pluggy==0.6.0
progressbar2==3.38.0
py==1.5.3
Pygments==2.2.0
pyGravatar==0.0.6
pymongo==3.6.1
pyramid==1.9.2
pyramid-chameleon==0.3
pyramid-debugtoolbar==4.4
pyramid-mako==1.0.2
pytest==3.6.1
pytest-cov==2.5.1
python-dateutil==2.7.3
python-utils==2.3.0
repoze.lru==0.7
six==1.11.0
translationstring==1.3
venusian==1.1.0
waitress==1.1.0
WebOb==1.8.1
WebTest==2.0.29
zope.deprecation==4.3.0
zope.interface==4.5.0
